moved attribute tests to its own source file
