+++ UnityScript2CSharp.Tests/UnityScript2CSharp.Tests.csproj
@@ -47,6 +47,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Tests.Attributes.cs" />
     <Compile Include="Tests.Operators.cs" />
     <Compile Include="Tests.Statements.Switch.cs" />
     <Compile Include="Tests.Statements.Yield.cs" />
